긴급수정

+++ BlackOlive/src/main/java/com/blackolive/app/domain/adminpage/ProductDTO.java
@@ -1,5 +1,7 @@
 package com.blackolive.app.domain.adminpage;
 
+import java.math.BigDecimal;
+
 import org.springframework.web.multipart.commons.CommonsMultipartFile;
 
 import lombok.AllArgsConstructor;
@@ -15,7 +17,7 @@ public class ProductDTO {
 	private String productDisplayId;
 	private String categorySmallId;
 	private String productName;
-	private int productPrice;
+	private BigDecimal productPrice;
 	private int productStock;
 	
 	private CommonsMultipartFile productImg;
